+++ src/main/java/com/microservice/user/UserApplication.java
@@ -3,13 +3,10 @@ package com.microservice.user;
 import org.modelmapper.ModelMapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
-@EnableJpaRepositories(basePackages = "com.microservice.user.dao")
-@EntityScan(basePackages = "com.microservice.entity")
+
 
 public class UserApplication {
