+++ src/main/java/com/microservice/user/dao/RegisterUserRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.microservice.entity.Users;
+import com.microservice.user.entity.Users;
 
 @Repository
 public interface RegisterUserRepository extends JpaRepository<Users, Long> {
