+++ src/main/java/com/microservice/user/service/LoginService.java
@@ -1,6 +1,6 @@
 package com.microservice.user.service;
 
-import com.microservice.entity.Users;
+import com.microservice.user.entity.Users;
 import com.microservice.user.exception.UserApplicationException;
 import com.microservice.user.request.UserLoginRequestDto;
