+++ src/main/java/com/microservice/user/service/RegistrationService.java
@@ -2,7 +2,7 @@ package com.microservice.user.service;
 
 import java.util.List;
 
-import com.microservice.entity.Users;
+import com.microservice.user.entity.Users;
 import com.microservice.user.exception.UserApplicationException;
 import com.microservice.user.request.UserRegistrationRequestDto;
