+++ src/main/java/com/microservice/user/service/impl/ResetPasswordServiceImpl.java
@@ -1,4 +1,4 @@
-package com.microservice.user.serviceImpl;
+package com.microservice.user.service.impl;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import com.microservice.entity.Users;
 import com.microservice.user.dao.RegisterUserRepository;
+import com.microservice.user.entity.Users;
 import com.microservice.user.exception.UserApplicationException;
 import com.microservice.user.request.ResetPasswordDto;
 import com.microservice.user.request.VerifyOtpDto;
